Increased sampling speed for pressure sensor
The MS5xxx library has the constants for configuring the oversampling rate, but does not expose any option to do so, and is hardcoded to the slowest one (far too slow for us).

I have included a lightly modified version of the library here, edited to include the ability to change oversampling rate. I changed the oversampling rate from 4096 to 1024, but may lower it further in the future.