Update Altimeter_Firmware.ino
I think it would be better if we did launch detection on an average velocity basis if we are just using altitude readings. This is just a basic format but still susceptible to erroneous readings. in future: take average value for pressure readings instead of just one (what is the best way to implement that?).