Added altitude calculation, changed library to MS5xxx
Created "calcAlt()" and "zeroAlt()" which ignore sea level pressures to give just the altitude difference between the pad and rocket while in flight. Fully moved to the MS5xxx library, but not 100% on the proper integration. I think the logic is actually usable but need to test on actual board. Also need to integrate averaging system on readings to improve resistance to spikes in temp/pressure.